MIB Parser: fix multi-component OID paths, remove search limit, strip DESCRIPTION false matches
Parser improvements:
- Handle IEEE-style multi-component OID assignments { org ieee(111) stds(2) ... }
  via new TryParseOidBraceContent() method (fixes LLDP-V2-MIB, LLDP-V2-TC-MIB)
- Remove hard 15000-char search limit for MODULE-IDENTITY blocks — use next
  definition boundary instead (fixes fsp150cm-entity 16KB, fsp150cm-facility 44KB)
- Strip quoted strings in CollectRawAssignments to prevent false matches inside
  DESCRIPTION blocks (fixes IF-MIB noTest/testCodeUnknown phantom definitions)
- Skip _synth_ nodes in validation (duplicate OID + unresolved parent checks)
- Add ieee802dot1mibs (1.3.111.2.802.1.1) to RootOids
- Change int to long for RangeMin/Max, SizeMin/Max (Unsigned32 overflow fix)
- Fix ModuleNameRegex to allow leading whitespace in MIB files
- Add SNMP-FRAMEWORK-MIB OIDs to RootOids

MIB Browser UI:
- Two-row header: buttons always visible, scrollable chips strip below
- Add Unload All MIBs button

Panel UI:
- Fix stray dot under device name when model/firmware empty
- Module tabs now wrap instead of hidden overflow
- Add Save Image feature (html2canvas)
- Client-side fixDependencyStatuses() safety net for validation

+++ SNMPSimMgr/Models/MibDefinition.cs
@@ -21,10 +21,10 @@ public class MibDefinition
 
     // Parsed syntax breakdown
     public string? BaseType { get; set; }
-    public int? RangeMin { get; set; }
-    public int? RangeMax { get; set; }
-    public int? SizeMin { get; set; }
-    public int? SizeMax { get; set; }
+    public long? RangeMin { get; set; }
+    public long? RangeMax { get; set; }
+    public long? SizeMin { get; set; }
+    public long? SizeMax { get; set; }
     public Dictionary<string, int>? EnumValues { get; set; }
 }
